requirements.txt
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2,<2
joblib>=1.3
streamlit>=1.32
fastapi>=0.110
uvicorn[standard]>=0.27
fpdf2>=2.7
matplotlib>=3.7
seaborn>=0.12